Switch from ansible-local to shell ansible
The ansible-local provisioner copies the playbook to a
directory under `/tmp` but the lockdown playbook we are using
ends up dumping the entire `/tmp` directory as it remediates things
and then ansible aborts.

So instead, lets run the local ansible provisioning using a simple
shell command out of `/app/ansible` instead.