Got ansible, packer, & terraform mostly working
This combines all the CIS baseline work with the Terraform work.

I'm getting a 404 on the Google object still but this has packer and
ansible and terraform all working happily together - now without having
to specify an explicit AMI id.

This also moves the scanning with openscap into an ansible role, making
all the provisioning after Ansible is installed run through ansible.

This has a preliminary script to rotate the instances in an ASG in it
also.