Initial work on re-splitting CodeDeploy / Packer
It looks like we are doing too much in Packer / Ansible and too little
in CodeDeploy with respect to setting up the Python spin application.

We should fix this up, this commit has the first stab at it, but it does
not have the emporor venv created the right way yet.

While this needs to get fixed, it doesn't have to happen before the
upcoming presentation on 2019-01-30.

See #16 for details