Quiet terraform provider version warnings
This fixes:
```
The following providers do not have any version constraints in configuration,
so the latest version was installed.

To prevent automatic upgrades to new major versions that may contain breaking
changes, it is recommended to add version = "..." constraints to the
corresponding provider blocks in configuration, with the constraint strings
suggested below.

* provider.aws: version = "~> 1.57"
* provider.google: version = "~> 1.20"
* provider.template: version = "~> 2.0"
```